requirements.txt
google.generativeai
datetime
pillow
os
